+++ src/shapes/mod.rs
@@ -20,10 +20,10 @@ pub enum Orientation {
 }
 
 pub trait Bounded {
-    fn bound_left(&self) -> f32;
-    fn bound_top(&self) -> f32;
-    fn bound_right(&self) -> f32;
-    fn bound_bottom(&self) -> f32;
+    fn bound_left(&self) -> f64;
+    fn bound_top(&self) -> f64;
+    fn bound_right(&self) -> f64;
+    fn bound_bottom(&self) -> f64;
 }
 
 pub trait RigidBody {
@@ -35,18 +35,18 @@ pub trait RigidBody {
 
 #[derive(Debug, Clone, Copy)]
 pub struct Particle {
-    pub mass: f32,
+    pub mass: f64,
     pub position: Vec2,
     pub velocity: Vec2,
     pub acceleration: Vec2,
 }
 
 impl Particle {
-    pub fn inverse_mass(&self) -> f32 {
+    pub fn inverse_mass(&self) -> f64 {
         if self.mass > 0.0 {
             1.0 / self.mass
         } else {
-            f32::INFINITY
+            f64::INFINITY
         }
     }
 }
@@ -58,7 +58,7 @@ pub trait ParticleLike {
 
 #[derive(Debug, Clone, Copy)]
 pub struct Material {
-    pub restitution: f32,
+    pub restitution: f64,
 }
 
 pub trait MaterialLike {
